requirements.txt
requests
Flask
python-dotenv
os